backend/requirements.txt
flask==3.0.0
werkzeug==3.0.1
numpy==1.26.2
pillow==10.1.0
trimesh==4.0.5
networkx==3.4.2
lxml==6.0.2
shapely==2.0.2
mapbox-earcut==1.0.1
scipy==1.11.4
gunicorn==21.2.0
flask-cors==4.0.0
